Split out files and add documentation
